+++ NEWS.md
@@ -1,3 +1,17 @@
+# watina 0.3.0
+
+#### New features
+
+- New function `cluster_locs()` to spatially cluster locations (well clusters) (#39)
+- More fun with `get_locs()`:
+    - optionally allow groundwater piezometers with missing `filterdepth` (argument `filterdepth_na`)
+    - optionally replace missing `filterdepth` values in a conservative way, based on `tubelength` (argument `filterdepth_guess`, #44)
+    - return eight more observation well attributes beside `filterdepth` (#44)
+    - four methods are available to aggregate observation well attributes per location (argument `obswell_aggr`, #44)
+- `convertdf_enc()` is now reexported from `inborutils`
+
+Further, a number of smaller fixes and enhancements were made.
+
 # watina 0.2.6
 
 - Redo fix `get_locs()`: calculation of `soilsurf_ost` (#43)
@@ -27,7 +41,7 @@ affecting the installation process on Windows (warnings were thrown).
 
 # watina 0.2.0
 
-## New features
+#### New features
 
 - Several functions have been added to query and process chemical data.
 - Three vignettes have been added.